helper_thread: split into separate file
It's nicely contained now, split it into its own set of files.
